[feat] info first section
