+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-mobile-datepicker": "^4.0.2",
     "react-router-dom": "^7.3.0",
+    "react-text-mask": "^5.5.0",
     "styled-components": "^6.1.15"
   },
   "devDependencies": {
