+++ src/styles/InfoPage/SecondSection.css
@@ -38,8 +38,9 @@
     height: 100%;
     display: flex;
     flex-direction: column;
-    margin-top: 50px;
-    gap: 20px;
+    margin-top: 80px;
+    gap: 10px;
+
 }
 
 .info-input-container {
@@ -77,24 +78,28 @@ input::placeholder {
     color: var(--gray03);
 }
 
-.next-button {
-    width: 100vw;
-    margin-left: -15px;
-    margin-top: auto;
-    margin-bottom: 10%;
+.nickname-check-button {
+    position: absolute;
+    right: 10px;  /* 컨테이너의 오른쪽에서 10px */
+    top: 50%;
+    transform: translateY(-50%);
     border: none;
-    height: 60px;
+    background-color: var(--main-coral);
     color: white;
     font-weight: var(--semibold);
-    font-size: 15px;
-}
-
-.next-button.active {
-    background-color: var(--main-coral); /* 원하는 색상 */
+    font-size: 13px;
+    width: 80px;
+    height: 30px;
+    padding: 5px 10px;
+    border-radius: 9px;
     cursor: pointer;
 }
-  
-.next-button.inactive {
-    background-color: var(--gray03); /* 비활성화 색상 */
-    cursor: not-allowed;
+
+.nickname-len {
+    width:100%;
+    text-align: right;
+    font-size: 11px;
+    font-weight: var(--medium);
+    color: var(--gray05);
+    padding: 5px;
 }
